pyproject.toml
[project]
name = "trading-stock-evaluator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.23.0",
    "scikit-learn>=1.4.0",
    "python-dotenv>=1.0.0",
    "psycopg[binary]>=3.3.0",
    "httpx>=1.1.0",
]
